Fixed printing wrong health when health is low

+++ allpixelmon.h
@@ -6,22 +6,22 @@
 
 typedef struct {
     char* name;
-    int dmg;
+    int8_t dmg;
 } fight_move;
 
 typedef struct {
     char *name;
     const uint8_t *bitmap;
-    int bitmap_width;
-    int bitmap_height;
+    int8_t bitmap_width;
+    int8_t bitmap_height;
     fight_move attacks[4];
 } pixelmon_type;
 
 typedef struct {
-    int pixelmon_id;
-    int health;
-    int level;
-    int xp;
+    int8_t pixelmon_id;
+    int8_t health;
+    int8_t level;
+    int8_t xp;
 } pixelmon;
 
 const int NUM_PIXELMON_TYPES = 2;
